+++ src/main/java/woowacourse/shoppingcart/application/CustomerService.java
@@ -5,13 +5,13 @@ import org.springframework.transaction.annotation.Transactional;
 import woowacourse.auth.application.AuthService;
 import woowacourse.shoppingcart.dao.CustomerDao;
 import woowacourse.shoppingcart.domain.customer.Customer;
-import woowacourse.shoppingcart.dto.CustomerUpdatePasswordRequest;
-import woowacourse.shoppingcart.dto.CustomerUpdateRequest;
-import woowacourse.shoppingcart.dto.TokenRequest;
-import woowacourse.shoppingcart.dto.CustomerResponse;
-import woowacourse.shoppingcart.dto.LoginRequest;
-import woowacourse.shoppingcart.dto.LoginResponse;
-import woowacourse.shoppingcart.dto.SignUpRequest;
+import woowacourse.shoppingcart.application.dto.request.CustomerUpdatePasswordRequest;
+import woowacourse.shoppingcart.application.dto.request.CustomerUpdateRequest;
+import woowacourse.shoppingcart.application.dto.request.TokenRequest;
+import woowacourse.shoppingcart.application.dto.response.CustomerResponse;
+import woowacourse.shoppingcart.application.dto.request.LoginRequest;
+import woowacourse.shoppingcart.application.dto.response.LoginResponse;
+import woowacourse.shoppingcart.application.dto.request.SignUpRequest;
 import woowacourse.shoppingcart.exception.datanotfound.CustomerDataNotFoundException;
 import woowacourse.shoppingcart.exception.datanotfound.LoginDataNotFoundException;
 import woowacourse.shoppingcart.exception.duplicateddata.CustomerDuplicatedDataException;
