+++ src/main/java/woowacourse/shoppingcart/ui/CartItemController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import woowacourse.shoppingcart.domain.Cart;
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.shoppingcart.dto.Request;
+import woowacourse.shoppingcart.application.dto.request.Request;
 import woowacourse.shoppingcart.application.CartService;
 
 import java.net.URI;
