+++ src/main/java/woowacourse/shoppingcart/ui/ProductController.java
@@ -5,7 +5,7 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.shoppingcart.dto.Request;
+import woowacourse.shoppingcart.application.dto.request.Request;
 import woowacourse.shoppingcart.application.ProductService;
 
 import java.net.URI;
