+++ src/test/java/woowacourse/shoppingcart/application/CustomerServiceTest.java
@@ -12,13 +12,13 @@ import org.springframework.test.context.jdbc.Sql;
 import org.springframework.transaction.annotation.Transactional;
 import woowacourse.auth.application.AuthService;
 import woowacourse.shoppingcart.dao.CustomerDao;
-import woowacourse.shoppingcart.dto.CustomerUpdatePasswordRequest;
-import woowacourse.shoppingcart.dto.CustomerUpdateRequest;
-import woowacourse.shoppingcart.dto.TokenRequest;
-import woowacourse.shoppingcart.dto.CustomerResponse;
-import woowacourse.shoppingcart.dto.LoginRequest;
-import woowacourse.shoppingcart.dto.LoginResponse;
-import woowacourse.shoppingcart.dto.SignUpRequest;
+import woowacourse.shoppingcart.application.dto.request.CustomerUpdatePasswordRequest;
+import woowacourse.shoppingcart.application.dto.request.CustomerUpdateRequest;
+import woowacourse.shoppingcart.application.dto.request.TokenRequest;
+import woowacourse.shoppingcart.application.dto.response.CustomerResponse;
+import woowacourse.shoppingcart.application.dto.request.LoginRequest;
+import woowacourse.shoppingcart.application.dto.response.LoginResponse;
+import woowacourse.shoppingcart.application.dto.request.SignUpRequest;
 import woowacourse.shoppingcart.exception.dataempty.CustomerDataEmptyException;
 import woowacourse.shoppingcart.exception.dataformat.CustomerDataFormatException;
 import woowacourse.shoppingcart.exception.datanotfound.CustomerDataNotFoundException;
